most of checkout api route written. Needs to send confirmations.
