+++ server/api/cart.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const {CartItem, Product} = require('../db/models')
+const { CartItem, Product, Order, OrderLine } = require('../db/models')
 
 //this sets up the cart if one does not exist.
 router.use((req, res, next) => {
@@ -28,7 +28,7 @@ GET: /api/cart/
 returns the cart
 */
 router.get('/', async (req, res, next) => {
-  if (!req.user){
+  if (!req.user) {
     res.json(req.session.cart)
   } else {
     const cartItems = await getCart(req.user.id)
@@ -50,12 +50,12 @@ example req.body is:
 router.post('/', async (req, res, next) => {
   //check if product exists
   const productData = await Product.findById(req.body.productId)
-  if (!productData){
+  if (!productData) {
     throw new Error('product of specified id does not exist')
   }
 
   //if not a user
-  if (!req.user){
+  if (!req.user) {
     //create a cartItemObject
     const cartItemObject = {
       productId: req.body.productId,
@@ -69,7 +69,7 @@ router.post('/', async (req, res, next) => {
     })
 
     //update or create cartItemObject in our session array
-    if (itemIdx !== -1){
+    if (itemIdx !== -1) {
       req.session.cart[itemIdx].quantity = cartItemObject.quantity
     } else {
       req.session.cart.push(cartItemObject)
@@ -92,7 +92,7 @@ router.post('/', async (req, res, next) => {
       cartItemData.dataValues.id :
       undefined
 
-      //upserts data into database
+    //upserts data into database
     await CartItem.upsert(
       {
         id: cartItemId,
@@ -117,38 +117,70 @@ example req.body is:
     productId : 3,
   }
 */
-router.delete('/', async (req, res, next) => {
-  //check if product exists
-  const productData = await Product.findById(req.body.productId)
-  if (!productData){
-    throw new Error('product of specified id does not exist')
-  }
 
+async function clearCartItem(req, productId) {
   //if not a user
-  if (!req.user){
+  if (!req.user) {
     //find index of the cart in our session
     const itemIdx = req.session.cart.findIndex((item) => {
-      return item.productId === req.body.productId
+      return item.productId === productId
     })
-
     //remove the item from the array(mutating method)
     req.session.cart.splice(itemIdx, 1)
-
-    //return the cart session
-    res.status(201).json(req.session.cart)
   } else {
-
     await CartItem.destroy({
       where: {
         userId: req.body.userId,
-        productId: req.body.productId,
+        productId,
       }
     })
     const cartItems = await getCart(req.user.id)
-
     req.session.cart = cartItems
-    res.status(201).json(cartItems)
   }
+  return req.session.cart
+}
+
+router.delete('/', async (req, res, next) => {
+  //check if product exists
+  const productData = await Product.findById(req.body.productId)
+  if (!productData) {
+    throw new Error('product of specified id does not exist')
+  }
+  const curCart = clearCartItem(req, req.body.productId)
+  res.status(201).json(curCart)
+})
+
+router.post('/cart/checkout', async (req, res, next) => {
+  const userId = req.user ? req.user.id : null
+  const orderEmail = req.user ? req.user.email : req.body.email
+  const totalAmount = req.body.cart.reduce((total, item) => {
+    total += item.price * item.quantity
+  })
+  const orderCart = req.body.cart.map(async (item) => {
+    const product = await Product.findById(item.productId)
+    return product.dataValues
+  })
+  const orderInfo = await Order.create({
+    userId,
+    orderEmail,
+    totalAmount,
+  })
+  await OrderLine.bulkCreate(orderCart.map(item => {
+    return {
+      orderId: orderInfo.dataValues.id,
+      productId: item.id,
+      title: item.title,
+      description: item.description,
+      price: item.price,
+      imageUrl: item.imageUrl,
+      quantity: item.quantity,
+    }
+  }))
+  req.body.cart.forEach(item => {
+    clearCartItem(req, item.id)
+  })
+  const cart = await getCart()
+  res.setStatus(201).json({ cart, orderInfo })
 })
 
 module.exports = router
